The MSE Learning Project. (Final goal is to gen Adaptive Sampling Map for Monte Carlo rendering)

+++ MathFunctions.h
@@ -8,6 +8,6 @@ void MatMultAB(cublasHandle_t handle, Matrix<float>& A, Matrix<float>& B, Matrix
 void MatMultATB(cublasHandle_t handle, Matrix<float>& A, Matrix<float>& B, Matrix<float>& C);
 void MatMultABT(cublasHandle_t handle, Matrix<float>& A, Matrix<float>& B, Matrix<float>& C);
 void SumRows(cublasHandle_t handle, Matrix<float>& A, Matrix<float>& B);
-void ComputeNorm2(cublasHandle_t handle, Matrix<float>& A, float& res);
+void ComputeNorm2(cublasHandle_t handle, Matrix<float>& A, double& res);
 
 #endif
